+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/QueryPlanConstraint.java
@@ -92,6 +92,11 @@ public class QueryPlanConstraint implements PlanHashable, PlanSerializable {
     }
 
     @Nonnull
+    public QueryPlanConstraint compose(@Nonnull final QueryPlanConstraint otherQueryPlanConstraint) {
+        return composeConstraints(ImmutableList.of(this, otherQueryPlanConstraint));
+    }
+
+    @Nonnull
     @Override
     public PQueryPlanConstraint toProto(@Nonnull final PlanSerializationContext serializationContext) {
         return PQueryPlanConstraint.newBuilder().setPredicate(predicate.toQueryPredicateProto(serializationContext)).build();
@@ -105,8 +110,8 @@ public class QueryPlanConstraint implements PlanHashable, PlanSerializable {
     }
 
     @Nonnull
-    public static QueryPlanConstraint compose(@Nonnull final Collection<QueryPlanConstraint> constraints) {
-        return new QueryPlanConstraint(AndPredicate.andOrTrue(constraints.stream().map(QueryPlanConstraint::getPredicate).collect(Collectors.toList())));
+    public static QueryPlanConstraint composeConstraints(@Nonnull final Collection<QueryPlanConstraint> constraints) {
+        return new QueryPlanConstraint(AndPredicate.and(constraints.stream().map(QueryPlanConstraint::getPredicate).collect(Collectors.toList())));
     }
 
     @Nonnull
@@ -116,7 +121,7 @@ public class QueryPlanConstraint implements PlanHashable, PlanSerializable {
 
     @Nonnull
     public static QueryPlanConstraint ofPredicates(@Nonnull final Collection<QueryPredicate> predicates) {
-        return new QueryPlanConstraint(AndPredicate.andOrTrue(predicates));
+        return new QueryPlanConstraint(AndPredicate.and(predicates));
     }
 
     @Nonnull
@@ -162,7 +167,7 @@ public class QueryPlanConstraint implements PlanHashable, PlanSerializable {
         public QueryPlanConstraint getConstraint(@Nonnull final RecordQueryPlan plan) {
             visit(plan);
             final var constraints = builder.build();
-            return QueryPlanConstraint.ofPredicate(AndPredicate.andOrTrue(constraints.stream().map(QueryPlanConstraint::getPredicate).collect(Collectors.toList())));
+            return QueryPlanConstraint.ofPredicate(AndPredicate.and(constraints.stream().map(QueryPlanConstraint::getPredicate).collect(Collectors.toList())));
         }
     }
 }