+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/predicates/DatabaseObjectDependenciesPredicate.java
@@ -36,6 +36,8 @@ import com.apple.foundationdb.record.RecordStoreState;
 import com.apple.foundationdb.record.metadata.Index;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreBase;
 import com.apple.foundationdb.record.query.plan.cascades.AliasMap;
+import com.apple.foundationdb.record.query.plan.cascades.BooleanWithConstraint;
+import com.apple.foundationdb.record.query.plan.cascades.ValueEquivalence;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryPlan;
 import com.google.auto.service.AutoService;
 import com.google.common.collect.ImmutableSet;
@@ -129,15 +131,16 @@ public class DatabaseObjectDependenciesPredicate extends AbstractQueryPredicate 
         return semanticEquals(other, AliasMap.identitiesFor(getCorrelatedTo()));
     }
 
+    @Nonnull
     @Override
-    public boolean equalsWithoutChildren(@Nonnull final QueryPredicate other, @Nonnull final AliasMap aliasMap) {
-        if (!super.equalsWithoutChildren(other, aliasMap)) {
-            return false;
-        }
-
-        final DatabaseObjectDependenciesPredicate otherDatabaseObjectDependenciesPredicate =
-                (DatabaseObjectDependenciesPredicate)other;
-        return usedIndexes.equals(otherDatabaseObjectDependenciesPredicate.usedIndexes);
+    public BooleanWithConstraint equalsWithoutChildren(@Nonnull final QueryPredicate other,
+                                                       @Nonnull final ValueEquivalence valueEquivalence) {
+        return super.equalsWithoutChildren(other, valueEquivalence)
+                .filter(ignored -> {
+                    final DatabaseObjectDependenciesPredicate otherDatabaseObjectDependenciesPredicate =
+                            (DatabaseObjectDependenciesPredicate)other;
+                    return usedIndexes.equals(otherDatabaseObjectDependenciesPredicate.usedIndexes);
+                });
     }
 
 
@@ -169,8 +172,8 @@ public class DatabaseObjectDependenciesPredicate extends AbstractQueryPredicate 
     }
 
     @Nonnull
-    public static <M extends Message> DatabaseObjectDependenciesPredicate fromPlan(@Nonnull final RecordMetaData recordMetaData,
-                                                                                   @Nonnull final RecordQueryPlan plan) {
+    public static DatabaseObjectDependenciesPredicate fromPlan(@Nonnull final RecordMetaData recordMetaData,
+                                                               @Nonnull final RecordQueryPlan plan) {
         final List<String> usedIndexesNamesList = Lists.newArrayList(plan.getUsedIndexes());
         // we have to do this to get a proper stable order
         Collections.sort(usedIndexesNamesList);