+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/values/Value.java
@@ -32,6 +32,7 @@ import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreBase;
 import com.apple.foundationdb.record.query.expressions.Comparisons;
 import com.apple.foundationdb.record.query.plan.cascades.AliasMap;
+import com.apple.foundationdb.record.query.plan.cascades.BooleanWithConstraint;
 import com.apple.foundationdb.record.query.plan.cascades.Correlated;
 import com.apple.foundationdb.record.query.plan.cascades.CorrelationIdentifier;
 import com.apple.foundationdb.record.query.plan.cascades.Formatter;
@@ -40,6 +41,9 @@ import com.apple.foundationdb.record.query.plan.cascades.Narrowable;
 import com.apple.foundationdb.record.query.plan.cascades.Quantifier;
 import com.apple.foundationdb.record.query.plan.cascades.ScalarTranslationVisitor;
 import com.apple.foundationdb.record.query.plan.cascades.TreeLike;
+import com.apple.foundationdb.record.query.plan.cascades.UsesValueEquivalence;
+import com.apple.foundationdb.record.query.plan.cascades.ValueEquivalence;
+import com.apple.foundationdb.record.query.plan.cascades.debug.Debugger;
 import com.apple.foundationdb.record.query.plan.cascades.expressions.RelationalExpression;
 import com.apple.foundationdb.record.query.plan.cascades.predicates.Placeholder;
 import com.apple.foundationdb.record.query.plan.cascades.predicates.PredicateWithValueAndRanges;
@@ -78,8 +82,9 @@ import java.util.stream.Stream;
 /**
  * A scalar value type.
  */
+@SuppressWarnings("OptionalUsedAsFieldOrParameterType")
 @API(API.Status.EXPERIMENTAL)
-public interface Value extends Correlated<Value>, TreeLike<Value>, PlanHashable, Typed, Narrowable<Value>, PlanSerializable {
+public interface Value extends Correlated<Value>, TreeLike<Value>, UsesValueEquivalence<Value>, PlanHashable, Typed, Narrowable<Value>, PlanSerializable {
 
     @Nonnull
     @Override
@@ -317,38 +322,96 @@ public interface Value extends Correlated<Value>, TreeLike<Value>, PlanHashable,
             return false;
         }
 
-        final Value otherValue = (Value)other;
-        if (!equalsWithoutChildren(otherValue, aliasMap)) {
-            return false;
-        }
-
-        final Iterator<? extends Value> children = getChildren().iterator();
-        final Iterator<? extends Value> otherChildren = otherValue.getChildren().iterator();
-
-        while (children.hasNext()) {
-            if (!otherChildren.hasNext()) {
-                return false;
-            }
-
-            if (!children.next().semanticEquals(otherChildren.next(), aliasMap)) {
-                return false;
-            }
-        }
-
-        return !otherChildren.hasNext();
+        return semanticEquals(other, ValueEquivalence.fromAliasMap(aliasMap)).isTrue();
     }
 
-    @SuppressWarnings({"unused", "PMD.CompareObjectsWithEquals"})
-    default boolean equalsWithoutChildren(@Nonnull final Value other,
-                                          @Nonnull final AliasMap equivalenceMap) {
+    /**
+     * Overriding method of {@link UsesValueEquivalence#semanticEquals(Object, ValueEquivalence)} that attempts to
+     * assert equivalence of {@code this} and {@code other} using the {@link ValueEquivalence} that was passed in.
+     * @param other the other object to compare this object to
+     * @param valueEquivalence the value equivalence
+     * @return a boolean monad {@link BooleanWithConstraint} that is either effectively {@code false} or {@code true}
+     *         under the assumption that a contained query plan constraint is satisfied
+     */
+    @Nonnull
+    @Override
+    @SuppressWarnings("PMD.CompareObjectsWithEquals")
+    default BooleanWithConstraint semanticEquals(@Nullable final Object other,
+                                                 @Nonnull final ValueEquivalence valueEquivalence) {
         if (this == other) {
-            return true;
+            return BooleanWithConstraint.alwaysTrue();
         }
 
-        return other.getClass() == getClass();
+        if (!(other instanceof Value)) {
+            return BooleanWithConstraint.falseValue();
+        }
+
+        final var thisOther = semanticEqualsTyped((Value)other, valueEquivalence);
+
+        Debugger.sanityCheck(() -> {
+            final var inverseValueEquivalenceMaybe = valueEquivalence.inverseMaybe();
+            Verify.verify(inverseValueEquivalenceMaybe.isPresent());
+            final var otherThis =
+                    ((Value)other).semanticEqualsTyped(this, inverseValueEquivalenceMaybe.get());
+            Verify.verify(thisOther.isTrue() == otherThis.isTrue());
+        });
+
+        if (thisOther.isFalse()) {
+            //
+            // By the looks of it, otherValue is not equal to this value. However, maybe it's already in the
+            // valueEquivalence.
+            //
+            return valueEquivalence.isDefinedEqual(this, (Value)other);
+        }
+
+        return thisOther;
     }
 
     @Nonnull
+    @Override
+    default BooleanWithConstraint semanticEqualsTyped(@Nonnull final Value other,
+                                                      @Nonnull final ValueEquivalence valueEquivalence) {
+        final var equalsWithoutChildren = equalsWithoutChildren(other);
+        if (equalsWithoutChildren.isFalse()) {
+            return BooleanWithConstraint.falseValue();
+        }
+
+        var constraint = equalsWithoutChildren;
+        final Iterator<? extends Value> children = getChildren().iterator();
+        final Iterator<? extends Value> otherChildren = other.getChildren().iterator();
+
+        while (children.hasNext()) {
+            if (!otherChildren.hasNext()) {
+                return BooleanWithConstraint.falseValue();
+            }
+
+            final var isChildEquals =
+                    children.next().semanticEquals(otherChildren.next(), valueEquivalence);
+            if (isChildEquals.isFalse()) {
+                return BooleanWithConstraint.falseValue();
+            }
+
+            constraint = constraint.composeWithOther(isChildEquals);
+        }
+
+        if (otherChildren.hasNext()) {
+            // otherValue has more children, it cannot be equivalent
+            return BooleanWithConstraint.falseValue();
+        }
+
+        return constraint;
+    }
+
+    @Nonnull
+    @SuppressWarnings({"unused", "PMD.CompareObjectsWithEquals"})
+    default BooleanWithConstraint equalsWithoutChildren(@Nonnull final Value other) {
+        if (this == other) {
+            return BooleanWithConstraint.alwaysTrue();
+        }
+
+        return other.getClass() == getClass() ? BooleanWithConstraint.alwaysTrue() : BooleanWithConstraint.falseValue();
+    }
+
     default boolean canResultInType(@Nonnull final Type type) {
         return false;
     }
@@ -483,6 +546,8 @@ public interface Value extends Correlated<Value>, TreeLike<Value>, PlanHashable,
                 .collect(ImmutableList.toImmutableList());
 
         final var correlatedToDifference = Sets.newHashSet(Sets.difference(getCorrelatedTo(), aliasMap.sources()));
+        // TODO This seems highly suspicious -- I think this should rather be the unions of all correlatedTos that needs
+        //      to be retained.
         correlatedTos.forEach(correlatedToDifference::retainAll);
 
         final var equivalenceMap = aliasMap.toBuilder()