+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/plans/RecordQueryInUnionPlan.java
@@ -389,6 +389,7 @@ public abstract class RecordQueryInUnionPlan implements RecordQueryPlanWithChild
      * @param inner the input/inner plan to this in-union
      * @param inSources a list of outer in-sources
      * @param comparisonKeyValues values by which the results of both plans are ordered
+     * @param isReverse indicator if {@code comparisonKeyValues} should be considered reversed (inverted).
      * @param maxNumberOfValuesAllowed maximum number of parallel legs of this in-union
      * @param internal indicator if bindings are modelled using correlation or old-style in-bindings
      * @return a new plan that will return the union of all results from both child plans
@@ -397,12 +398,13 @@ public abstract class RecordQueryInUnionPlan implements RecordQueryPlanWithChild
     public static RecordQueryInUnionOnValuesPlan from(@Nonnull final Quantifier.Physical inner,
                                                       @Nonnull final List<? extends InSource> inSources,
                                                       @Nonnull final List<? extends Value> comparisonKeyValues,
+                                                      final boolean isReverse,
                                                       final int maxNumberOfValuesAllowed,
                                                       @Nonnull final Bindings.Internal internal) {
         return RecordQueryInUnionOnValuesPlan.inUnion(inner,
                 inSources,
                 comparisonKeyValues,
-                Quantifiers.isReversed(ImmutableList.of(inner)),
+                isReverse,
                 maxNumberOfValuesAllowed,
                 internal);
     }