+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/query/plan/match/FilterMatcher.java
@@ -45,7 +45,7 @@ public class FilterMatcher extends PlanMatcherWithChild {
     public boolean matchesSafely(@Nonnull RecordQueryPlan plan) {
         final QueryPredicate predicate;
         if (plan instanceof RecordQueryPredicatesFilterPlan) {
-            predicate = AndPredicate.andOrTrue(((RecordQueryPredicatesFilterPlan)plan).getPredicates());
+            predicate = AndPredicate.and(((RecordQueryPredicatesFilterPlan)plan).getPredicates());
         } else {
             return false;
         }